iio: ad9680: Add basic support for ad9694
The AD9694 is a four channel 1.25 GSPS ADC with a JESD204 interface. Its
register map is quite similar to the AD9680 and so it can be supported by
the same driver.

Notable differences are that the AD9694 has a per channel input range
configuration whereas on the AD9680 all channels share the same input
range.

In addition the AD9694 has a special register write sequence that need to
be performed when bringing up the JESD204 link to ensure proper operation.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
(cherry picked from commit 5acae0fec1e32a633aa5b9bd7fbc079a26541820)