iio: Move attach/detach of the poll func to the core (part2)
This amends and includes commit eee97d12665fef8c ("iio: Move attach/detach
of the poll func to the core").
A lot of drivers were updated since this patch was added and we need to
update all of them until a consensus is reached upstream about whether to
do this patch (upstream) or not (in which case we will need to revert it).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>