fix: hwmon: axi_fan_control: Set tacho parameters
The driver was writting on the wrong register to set the tacho
parameters. For the core to start monitoring the signal, the driver has
to set the TACH_TOLERANCE register. This has to be done after writting
on TACH_PERIOD.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 0e51d44db536ea074239f97811e72112227df093)